+++ src/lib/components/VideoView.svelte
@@ -1,46 +1,25 @@
 <script lang="ts">
+	import VideoPlayer from '$lib/components/VideoPlayer.svelte';
+
 	import { selectedVideo } from '$lib/stores/selectedVideo';
-	import { appWindow } from '@tauri-apps/api/window';
-	import { tauri } from '@tauri-apps/api';
 
-	let fullscreen = false;
-	let player: any;
-	let videoPath = !$selectedVideo ? '' : tauri.convertFileSrc($selectedVideo);
+	let name: string = '';
 
-	function toggleFullscreen() {
-		console.log('toggleFullscreen');
-		fullscreen = !fullscreen;
-		appWindow.setFullscreen(fullscreen);
-	}
-
-	async function initializePLayer() {
-		if (!$selectedVideo) return;
-
-		let videoPath = tauri.convertFileSrc($selectedVideo);
-
-		// @ts-ignore
-		player = new window.Playerjs({
-			id: 'player',
-			file: videoPath,
-			autoplay: false
-		});
-	}
-
-	async function reloadPlayer() {
-		if (player) player.api('file', videoPath, '*');
-		else initializePLayer();
+	function truncateText(text: string) {
+		return text.split('\\').at(-1)!.length > 60
+			? text.split('\\').at(-1)!.substring(0, 60) + '[...]'
+			: text.split('\\').at(-1)!;
 	}
 
 	$: if ($selectedVideo) {
-		videoPath = tauri.convertFileSrc($selectedVideo);
-		reloadPlayer();
+		name = truncateText($selectedVideo);
 	}
 </script>
 
-<svelte:head>
-	<script defer src="/playerjs.js" type="text/javascript" on:load={initializePLayer}></script>
-</svelte:head>
-
-<!-- Video player -->
-<section class="basis-1/3" id="player" on:fullscreenchange={toggleFullscreen} />
-<!-- Video player -->
+<section class="basis-1/3 flex flex-col">
+	<VideoPlayer />
+	<section>
+		<p class="text-2xl font-bold">{name}</p>
+		<p class="text-base" />
+	</section>
+</section>
